launch and landing scene merged
launch scene now switches out the loading bay for a different prefab when in "landing mode"
landing mode is determined by the "did we warp in" flag on the jump system. If its true then the ship is placed at the jumpexit point and the landing bay is loaded. If not then the ship is put in the default launch bay